Taiping Heavenly Kingdom texts, prelim var character analysis
The Taiping Heavenly Kingdom outputted more than several Wenyan texts in its 14-year days, and so it has been added to the corpus. Its nature as a country is a bit weird - it indeed had its own culture, government, and so on, but is most commonly framed as a rebellion and is a "blip" in the time of the Qing. I decided to place it in the China section given the geographic placement but do use its proper name as a compromise.